+++ Core/Init.lua
@@ -374,7 +374,7 @@ local function DetermineML()
     if IsInRaid() then
         for i = 1, GetNumGroupMembers() do
             local name, rank = GetRaidRosterInfo(i)
-            if rank == 2 then -- Raid leader
+            if rank == 2 and not LoothingUtils.IsSecretValue(name) then -- Raid leader
                 return LoothingUtils.NormalizeName(name)
             end
         end
@@ -386,6 +386,7 @@ local function DetermineML()
             local unit = "party" .. i
             if UnitExists(unit) and UnitIsGroupLeader(unit) then
                 local name, realm = UnitName(unit)
+                if LoothingUtils.IsSecretValue(name) then return nil end
                 if realm and realm ~= "" then
                     return name .. "-" .. realm
                 end
@@ -1408,16 +1409,16 @@ Loothing.debug = false
 
 function Loothing:Debug(...)
     if self.debug then
-        print("|cff00ff00[Loothing Debug]|r", ...)
+        print("|cff00ff00[Loothing Debug]|r", LoothingUtils.SecretsForPrint(...))
     end
 end
 
 function Loothing:Error(...)
-    print("|cffff0000[Loothing Error]|r", ...)
+    print("|cffff0000[Loothing Error]|r", LoothingUtils.SecretsForPrint(...))
 end
 
 function Loothing:Print(...)
-    print("|cff00ccff[Loothing]|r", ...)
+    print("|cff00ccff[Loothing]|r", LoothingUtils.SecretsForPrint(...))
 end
 
 --[[--------------------------------------------------------------------